+++ tests/testthat.R
@@ -1,0 +1,6 @@
+library(testthat)
+library(aae.db)
+library(dplyr)
+library(tidyr)
+
+test_check("aae.db")